# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: OneQR Business Growth Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
        subscriptionPlan: { type: string, enum: [free, pro, business] }
        subscriptionStatus: { type: string, enum: [active, expired, cancelled, trial] }
        subscriptionExpiresAt: { type: string, format: date-time, nullable: true }
        isAdmin: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, email, name, subscriptionPlan, subscriptionStatus, isAdmin, createdAt]
...